+++ app/controllers/manage_owners_controller.rb
@@ -6,16 +6,27 @@ class ManageOwnersController < ApplicationController
     end  
     def edit
         id = params[:id]
-        @owner = User.find(id)
-        render "edit_owner"
+        if User.owners.exists?(id)
+            @owner = User.find(id)
+            render "edit_owner"
+        else  
+            flash[:error] = "Invalid"
+            redirect_to manage_owners_path
+        end      
     end    
     
     def update
         id = params[:id]
-        owner = User.find(id)
-        owner.name = params[:name]
-        owner.email = params[:email]
-        owner.save
+        if User.owners.exists?(id)
+            owner = User.find(id)
+            owner.name = params[:name]
+            owner.email = params[:email]
+            if owner.save
+                flash[:success] ="Updated"
+            end    
+        else  
+            flash[:error] = "Invalid"  
+        end    
         redirect_to manage_owners_path
       
     end 
@@ -23,18 +34,27 @@ class ManageOwnersController < ApplicationController
         render "new_owner"
     end 
     def create
-        User.create!(
+        user = User.new(
             name: params[:name],
             email: params[:email],
             role: "owner",
             password: params[:password]
         )
+        if user.save
+            flash[:success] = "Added Successfully"
+        else  
+            flash[:error] = "Invalid"  
+        end    
         redirect_to manage_owners_path
     end  
     def destroy
         id = params[:id]
-        owner = User.find(id)
-        owner.destroy
+        if User.owners.exists?(id)    
+            owner = User.find(id)
+            owner.destroy
+        else
+            flash[:error] = "Some Error Occured.Try Again" 
+        end      
         redirect_to manage_owners_path
     end            
 end
